REFACTORED: add classnames in heroesListItem; add separated files for selectors
