+++ 05_fifthApp/src/components/app/App.js
@@ -5,8 +5,8 @@ import { useHttp } from "../../hooks/http.hook";
 import { HeroesList } from "../heroesList/HeroesList";
 import { HeroesFilters } from "../heroesFilters/HeroesFilters";
 import { HeroesAddForm } from "../heroesAddForm/HeroesAddForm";
-import { fetchFiltersThunk } from "../../actions";
-import { fetchHeroesThunk } from "../../actions";
+import { fetchFiltersThunk } from "../../store/actions";
+import { fetchHeroesThunk } from "../../store/actions";
 import "./app.scss";
 
 export const App = () => {
