+++ 05_fifthApp/src/store/index.js
@@ -1,12 +1,11 @@
 import { legacy_createStore as createStore, combineReducers, applyMiddleware } from "redux";
-import { heroes } from "../reducers/heroes";
-import { filters } from "../reducers/filters";
+import { heroes } from "../store/reducers/heroes";
+import { filters } from "../store/reducers/filters";
 import ReduxThunk from 'redux-thunk';
 
 const store = createStore(  
   combineReducers({heroes, filters}),
   applyMiddleware(ReduxThunk),
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 );
 
 export default store;
